requirements.txt
flask
pandas
werkzeug
python-dotenv
scikit-learn
matplotlib
seaborn
joblib